Move SGX signing code into the types module
Since we have combined sgx-types and sgx-crypto into a single crate, the
artificial extension trait is no longer needed. Instead, we just implement
`sign()` on `Measurement` directly.

Additionally, remove the dependency on the patched version of openssl. If
upstream decides to merge my patches, then we can rebase on top of those. Until
then, we'll use the current inefficient conversions that upstream has.