Refer to correct struct and kernel header
Just for clarity's sake since indeed both structs and both headers are
relevant to this crate :-)

This file corresponds to ioctls that are called only on /dev/sev,
meaning the correct header and struct are sev_issue_cmd and psp-sev.h,
respectively.

A different module will be responsible for dealing with the SEV ioctls
that are called on /dev/kvm. Those ioctls will use struct kvm_sev_cmd
and include/uapi/linux/kvm.h.