Remove just-in-time certificate signing
The only situation in which this is useful is if the certificate chain
doesn't exist in the first place. This doesn't buy us anything if the
certificate chain exists but is invalid (i.e., it was exported and then
someone ran a platform reset).

Before, running the test suite would build 150 crates. Now, it builds
61.

Signed-off-by: Connor Kuehl <ckuehl@redhat.com>