Re-shape PlatformStatus into sev_user_data_status
This struct must take the form of the sev_user_data_status struct that the
AMD SEV kernel module exports. Right now, it resembles the command buffer
that the kernel is meant to send to the firmware; which is wrong because
the struct exported by the kernel and the command buffer that the firmware
expects are different. This results in data loss (most recently observed in
losing information in the flags variable).

For example, when running `sevctl show flags` on Rome (without this patch):

[connorkuehl@rome sevctl]$ cargo run -- show flags
   Compiling sevctl v0.1.0 (/home/connorkuehl/git/sevctl)
    Finished dev [unoptimized + debuginfo] target(s) in 2.16s
     Running `target/debug/sevctl show flags`
owned

^^^
Note that the "es" flag we expect to see is missing.

With this patch:

[connorkuehl@rome sevctl]$ cargo run -- show flags
   Compiling sev v0.1.0 (/home/connorkuehl/git/sev)
   Compiling sevctl v0.1.0 (/home/connorkuehl/git/sevctl)
    Finished dev [unoptimized + debuginfo] target(s) in 3.27s
     Running `target/debug/sevctl show flags`
owned
es

Signed-off-by: Connor Kuehl <ckuehl@redhat.com>