session: allow verifying an external digest
In some cases, such as when the digest of each launch measurement is
pre-generated beforehand, it's useful to be able to verify an
externally generated digest instead of having to pass the payload data
through Session<Measuring>::update_data().

This change implements Session<Measuring>::verify_with_digest(), which
behaves in the same way as Session<Measuring>::verify() but verifying
a digest that is provided as an argument to the method.

Signed-off-by: Sergio Lopez <slp@redhat.com>