host: Adding better error handling
From the changes suggested in the PR, this adds in a much more
user-friendly error handling when invalid GUIDs provided by
the user are encountered. Instead of using an assert, we bubble up a
custom SnpCertError which is wrapped in a UserApiError. The wrapper
covers Firmware, IO, and User API related errors.

Signed-off-by: Larry Dewey <larry.dewey@amd.com>