63: Patching for userspace breakage.
The kernel patch by pgonda@google.com in kernel hash 47894e0f changed the ioctl return for SNP_GET_EXT_REPORT to succeed instead of returning an error when encountering an invalid certificate length. This was done to keep the cryptography safe. Because of this, any kernel before this hash (5.19, etc) is broken, and we need to patch the way the API handles the error response.

resolves: #63

Signed-off-by: Larry Dewey <larry.dewey@amd.com>