ffi: Add capi feature and C APIs for SEV{_ES}_INIT
The capi feature is searched for with the cargo-c commands (cbuild,
ctest, cinstall, etc). It includes the C FFI to the SEV API.

cargo-c can generate and install a correct pkg-config file, a static library and
a dynamic library, and a C header to be used by any C (and C-compatible)
software.

Each INIT API takes three paramaters:
* The encrypting VM's file descriptor.
* The file descriptor of /dev/sev.
* A C int placeholder for a firmware error code in the event of an API
  failure.

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>