ext_config: Fixes config only operations.
Previously we were passing the u64 value of the
pointer to the vector of bytes the extended
report certificates were being loaded from.
However, when handling the case of an extended
configuration only, because the pointer value
of an empty vector is not 0, this was causing
an error in the kernel when attempting to access
the "pointer" being provided. By modifying the
underlying type of `bytes` to `Option<Vec<u8>>`,
it allows the default value to be set to `None`,
and conditionally update the extended config
`certs_address` field when certificate bytes are
present.

Also updated the TryFrom<ExtConfig> for
SnpExtSetConfig to use the reference instead of
a copy. This prevents the underlying config
object from being dropped after the if-let block.

Signed-off-by: Larry Dewey <larry.dewey@amd.com>