Updated OVMF to add new section and updated tests with new OVMF binaries
New OVMF changes added section SVSM_CAA to OvmfSevMetadata.

To calculate the measurement with this new section, the new section
has to be added and needs to be treated as zero pages when encountered.

Upstream OVMF has these changes for AmdSevOvmf, so both OVMF test
binaries AmdSevOvmf and OvmfX64 were regenerated to match the latest EDk2 release:
https://github.com/tianocore/edk2/releases/tag/edk2-stable202405

The tests were updated to reflect the new measurements with the changes
done to OVMF.

Signed-off-by: DGonzalezVillal <Diego.GonzalezVillalobos@amd.com>