rand: Removing use of openssl rand for crypto
According to the documentation, openssl's `rand_bytes(...)` function will be
seeded with `/dev/urandom`, which is not sound for cryptographic
application. We will be switching to using the `rdrand` crate which will
utilize the X86_64 microcode call to `RDRAND`, instead. All X86_64
instructions are protected by the Guest Hypervisor Communication Block
(GHCB), and may thus be presumed secure for cryptographic practices in
the context of confidential compute.

Obviously this introduces a change to the API, requiring a new major
release of the crate.

Resolves: #189

Signed-off-by: Larry Dewey <larry.dewey@amd.com>