Adding the KeyInfo field into the attestation report
Since version 2 of the attestation report, instead of having just the
AUTHOR_KEY_EN fIELD in the attestation report, the report now offers
more data around the signature and the key being used in the attestation
report. This PR extends that field and makes it accessible in the
attestation report.

Signed-off-by: DGonzalezVillal <Diego.GonzalezVillalobos@amd.com>