firmware/snp: Update PlatformStatus bitflag fields
Initially, byte 0x3 of STRUCT_PLATFORM_STATUS only used bit 0 to
indicate if the RMP was initialized. Updates were made to include
bitfields to indicate if alias detection has completed since the
last system reset and there are no aliasing addresses (bit 1), as well
as if SEV-TIO is enabled (bit 3).

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>