Adding AttestationReport Verification with Certificate only
We currently only offer report verification with a full chain being
provided. Although it is safest to provide a whole chain to verify the
whole root of trust, there are cases people might want to divide the process
into individual steps. This allows users to verify the report with what
should be an already verified versioned endorsement key.

I also renamed the VCEK to VEK since we no offer support for both the
VLEK and VCEK.

Signed-off-by: DGonzalezVillal <Diego.GonzalezVillalobos@amd.com>