Attestation Report versioning Update
In spec 1.56 of the SEV firmware a new version of the attestation report
was introduced.

Here we are introducing a way to version the attestation report that
keeps security and backwards compatibility.

The main AttestationReport is now an enum that will contain the
different versions of the attestation report. This will not only handle
both of the Attestation reports, but it will also work as an interface.
Users will be able to use the enum to get any desired field and display
the report without having to manually unwrap the report themselves.

There are 2 new structs for the Attestation Report, one for each version.
There is a new trait called Attestable that all the attestation reports will implement, this
will allow users to attest their report regardless of the version.

The ReportRsp will now contain raw bytes, rather than the Attestation
Report Strucutre. The AttestationReport Enum has a TryFrom bytes that
will return the appropriate attestation report version according to the
first 4 bytes of the raw data.

Structs consumed by the attestation report that now have new fields
depending on the version, are now also versioned, and each report
will consume the appropriate version of that struct (look at PlatInfo).

We also add the sealed module with the sealed trait. This allows us to
seal traits we want people to be able to use, but not to be able to impl
themselves. For example in this PR we are now sealing the new trait
Attestable.

Signed-off-by: DGonzalezVillal <Diego.GonzalezVillalobos@amd.com>