firmware/host: fix legacy attestation report verification
This commit fixes legacy attestation report verification by addressing
a few problems:

1. Fix typo of `launch_digest` definition for the
`LegacyAttestationReport` structure: should be `DIGEST_SIZE` not
`POLICY_SIZE`.

2. Signature verification should be performed on the SHA256 digest,
not on the raw data. Additionally, the return value of `sig.verify()`
should be properly converted to the `Result` type.

3. The `EcdsaSignature` structure has a size of 512 bytes, but the
`LegacyAttestationReport` defines the signature as 144 bytes. Change
the type of the signature to an `Array<u8, 144>` with an appropriate
conversion trait.

Signed-off-by: Roman Penyaev <r.peniaev@gmail.com>