Refactor: Simplify Report Parsing by using Generation; Fix redundant errors
We have variants for each attestation report version (pre-turin and post-turin). As we add new versions the amount of variants we will need for each report will grow quickly and will make logic to handle the variants annoying and redundant. With this change we can make use of the already established GENERATION enum and it's logic to tell the proc generation atparsing time. Then we can make better use of the write_tcb and parse_tcb functions to parse according to the proc generation instead of a boolean. The only difference (so far) between version variants is the TCB format according to it's proc generation. This change will allow us to maintain a single variant per report version.

Also some lint errors were addressed to remove redundant calls of
io::Error::new.

Signed-off-by: DGonzalezVillal <Diego.GonzalezVillalobos@amd.com>