firmware/snp: Support custom Wrapped VLEK hashsticks in SNP_VLEK_LOAD
This commit updates the snp_vlek_load logic to support the new WrappedVLEKHashstick format introduced in the SNP firmware specification. Users can now provide raw hashstick bytes to load a custom VLEK hashstick.

The hashstick structure is passed to the snp_vlek_load command, and the FFI definition is still used to invoke the SNP_VLEK_LOAD IOCTL, as it requires a pointer to a buffer containing the serialized bytes.

This change allows for more flexibility and testing of VLEK workflows without relying solely on the KDS-provided values.

Signed-off-by: DGonzalezVillal <Diego.GonzalezVillalobos@amd.com>