Modifying SNP_VLEK_LOAD logic
The Wrapped VLEK hashstick was added to the SNP fw spec. The user can
now create their own wrapped hashstick by providing the raw hashstick
bytes. The hashstick structure is  now provided to the snp_vlek_load
command. We still use the FFI definition when callning SNP_VLEK_LOAD,
since we need to provide the address to a buffer containing the
bytes with the hashstick to the IOCTL command.

Signed-off-by: DGonzalezVillal <Diego.GonzalezVillalobos@amd.com>