Fix: Update Commit structures to accurately reflect TCBs
When we updated the TCB structure, we overlooked modifying the
commit-related structures to account for differences in TCB layout
across CPU generations.
This commit ensures that raw TCB bytes are correctly constructed for
the FFI layer, and that conversions from the user-level structure to the
FFI format use the appropriate TCB structure based on the processor
generation.
The API snp_set_config function remains the same, it will tell the CPU
generation using a CPUID call.

Signed-off-by: DGonzalezVillal <Diego.GonzalezVillalobos@amd.com>