snp: skip calling CPUID on non-x86 platform
When this crate is used to parse SNP hardware evidence on a non-x86
platform, it will still try to build with cpuid from the x86 crate,
which will not be found.

Instead, put cpuid behind a platform config flag. If we are not on x86,
don't call cpuid and return an error.

Signed-off-by: Tobin Feldman-Fitzthum <tobin@ibm.com>