Lax parser feature
The parser strictly checks reserved areas to be zero (even though the specification does not even require that for all of them). This is an issue when parsing newer versions of attestation reports where those areas might not be reserved anymore. The `lax-parser` feature disables that check, currently with some tradeoffs:

* No `WriteExt::skip_bytes`
* No `Encoder` impl for `AttestationReport`, `Signature`, and `WrappedVlekHashstick`
* No compatibility with `openssl` and `crypto_nossl`

Closes #343

Signed-off-by: Christopher Schramm <git@cschramm.eu>